+++ R/bf_corr_test.R
@@ -47,7 +47,7 @@ bf_corr_test <- function(data,
                          y,
                          bf.prior = 0.707,
                          caption = NULL,
-                         output = "results",
+                         output = "dataframe",
                          k = 2L,
                          ...) {
 
@@ -62,7 +62,7 @@ bf_corr_test <- function(data,
   # return the text results or the dataframe with results
   switch(
     EXPR = output,
-    "results" = bf_extractor(bf_object),
+    "dataframe" = bf_extractor(bf_object),
     bf_expr(bf_object, k = k, output = output, caption = caption, ...)
   )
 }